Proyecto base Eplicacion escritorio
Proyecto base para la funcion basica

+++ Grupo-1-DI/Base_Datos/Departamentos.cs
@@ -8,10 +8,10 @@ namespace Grupo_1_DI.Base_Datos
 {
     public class Departamentos
     {
-        public int id {  get; set; }
+        public long id {  get; set; }
         public string cod {  get; set; }
         public string nombre { get; set; }
         public bool activo { get; set; }
-        public int jefedep_id { get; set; }
+        public long jefedep_id { get; set; }
     }
 }
